CRUD taches + modif date en string

+++ src/main/java/com/cesi/projetJavaSpringBoot/repositories/TacheRepository.java
@@ -3,5 +3,5 @@ package com.cesi.projetJavaSpringBoot.repositories;
 import com.cesi.projetJavaSpringBoot.entities.Taches;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TacheRepository extends JpaRepository<Taches, Integer> {
+public interface TacheRepository extends JpaRepository<Taches, Long> {
 }
